move apply_history to mod replay
